pyproject.toml
[project]
name = "friday-tony-stark-demo"
version = "0.1.0"
description = "Friday — A Tony Stark-inspired MCP server"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fastmcp",
    "httpx",
    "livekit-agents[deepgram,groq,openai,sarvam,silero]>=1.5.1",
    "livekit-plugins-google>=1.5.1",
    "livekit-plugins-anthropic>=0.1.0",
    "python-dotenv",
    "anthropic",
    "python-pptx",
    "yfinance",
    "pvporcupine>=3.0.0",
    "pyaudio>=0.2.14",
]

[project.scripts]
friday = "server:main"
friday_voice = "agent_friday:dev" # Points to the wrapper

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["server.py"]
